Añadiendo rutas para perfiles de usuarios. Reunificando funcionalidad de themas en ProfileController. Personalizando mensajes de error.
