+++ routes/web.php
@@ -1,6 +1,6 @@
 <?php
 
-use App\Http\Controllers\appThemeController;
+use App\Http\Controllers\ProfileController;
 use App\Http\Controllers\Auth\GitHubLoginController;
 use App\Http\Controllers\HelpController;
 use Illuminate\Support\Facades\Route;
@@ -18,14 +18,18 @@ use Illuminate\Support\Facades\Route;
 
 // Guest
 
-Route::get('/', [appThemeController::class, 'initialConfig'])->name('home');
+Route::get('/', [ProfileController::class, 'initialConfig'])->name('home');
 Route::get('/privacy-policy', fn() => view('policy'))->name('privacy-policy');
 Route::get('/terms-of-service', fn() => view('terms'))->name('terms-service');
 Route::get('/help', [HelpController::class, 'index'])->name('help');
 
+Route::fallback(function() {
+    return view('errors.404');
+});
+
 // Layout Change Theme
 
-Route::post('/save-theme', [appThemeController::class, 'saveModifiedTheme'])
+Route::post('/save-theme', [ProfileController::class, 'saveModifiedTheme'])
     ->name('save-theme');
 
 // GitHub Login
@@ -36,18 +40,30 @@ Route::get('/login/github/callback', [GitHubLoginController::class, 'handleProvi
 
 // Auth User
 
-Route::middleware([
+Route::prefix('user')->middleware([
     'auth:sanctum',
     config('jetstream.auth_session'),
     'verified'
 ])->group(function () {
-    Route::get('/user/dashboard', [appThemeController::class, 'authUserThemeConfig'])
+
+    Route::get('/dashboard', [ProfileController::class, 'authUserThemeConfig'])
         ->name('dashboard');
 
-    Route::get('/user', function() {
+    Route::get('/', function() {
         return redirect()->route('dashboard');
     });
 
-    Route::get('/user/profile/sync/github', [GitHubLoginController::class, 'redirectToProvider'])
+    // GitHub User Account Sync
+    Route::get('/profile/sync/github', [GitHubLoginController::class, 'redirectToProvider'])
         ->name('github.sync');
+});
+
+Route::prefix('users')->middleware([
+    'auth:sanctum',
+    config('jetstream.auth_session'),
+    'verified'
+])->group(function () {
+
+    Route::get('/{slug}', [ProfileController::class, 'showProfilesMainPage'])
+        ->name('users.main');
 });
